# secp256k1: Use field val for y coord decompression.