@davecgh 2020-02-20
> Commit msg has "... still relies on **then** ..."

Fixed.